requirements.txt
flask==3.1.0
werkzeug==3.1.3
Pillow==10.2.0
ffmpeg-python==0.2.0
gunicorn==21.2.0